Add PageTreeInput.tsx for arrays of page references
